requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
asn1crypto==0.24.0
async-timeout==4.0.0
attrs==21.2.0
autobahn==19.8.1
Automat==0.7.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.0
certifi==2019.6.16
cffi==1.12.3
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.2
cryptography==36.0.0
daphne==3.0.2
decorator==5.1.0
defusedxml==0.6.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-appconf==1.0.3
django-bower==5.2.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-compressor==2.4.1
django-compressor-toolkit==0.6.0
django-constance==2.8.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-debug-toolbar==2.0
django-filter==2.4.0
django-js-reverse==0.9.1
django-libsass==0.7
django-macros==0.4.0
django-picklefield==2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==9.8.3
hiredis==2.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.29.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.1
kombu==5.2.1
libsass==0.19.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mixer==7.2.0
more-itertools==8.8.0
msgpack==1.0.2
msgpack-python==0.5.6
oauthlib==3.1.0
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycrypto==2.6.1
Pygments==2.10.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-cov==3.0.0
pytest-django==4.4.0
python-crontab==2.5.1
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2020.1
rcssmin==1.0.6
redis==4.0.2
requests==2.27.1
requests-oauthlib==1.2.0
rjsmin==1.1.0
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
rules==2.1
service-identity==21.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.2
traitlets==5.1.1
Twisted==19.7.0
txaio==18.8.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==4.6.0
git+https://github.com/jabguru/django-channels-presence.git@29e76e370c542ae5056fa5af3266721aca1136f9
responses==0.17.0